Only one build is necessary for all redirect tests.
Change has been suggested by @kwpolska in https://github.com/getnikola/nikola/pull/3317.
